requirements.txt
flask
aiohttp
requests
pycryptodome
